+++ src/app/app-routing.module.ts
@@ -1,4 +1,4 @@
-import { RegisterComponent } from './welcome/register/register.component';
+import { RegisterComponent } from './welcome/waitlist/register/register.component';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
 import { WelcomeComponent } from './welcome/welcome.component';
 import { NgModule } from '@angular/core';
