+++ src/app/welcome/waitlist/register/register.component.html
